+++ src/test/java/lotto/domain/MoneyTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static lotto.domain.LottoNumber.GAME_PRICE;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -12,9 +14,9 @@ public class MoneyTest {
     @Test
     @DisplayName("입력 금액 생성 테스트")
     public void money() {
-        Money money = new Money(1000);
+        Money money = new Money(BigDecimal.valueOf(1000));
 
-        Money expected = new Money(1000);
+        Money expected = new Money(BigDecimal.valueOf(1000));
 
         Assertions.assertThat(money).isEqualTo(expected);
     }
@@ -23,7 +25,7 @@ public class MoneyTest {
     @DisplayName("입력 금액 테스트(0원)")
     public void money2() {
         assertThatThrownBy(() -> {
-            new Money(0);
+            new Money(BigDecimal.valueOf(0));
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("[ERROR]");
     }
@@ -31,7 +33,7 @@ public class MoneyTest {
     @Test
     @DisplayName("구매 회수 테스트")
     public void money3() {
-        Money money = new Money(1000);
+        Money money = new Money(BigDecimal.valueOf(1000));
 
         int purchaseCount = money.getPurchaseCount(GAME_PRICE);
 
@@ -42,7 +44,7 @@ public class MoneyTest {
     @DisplayName("입력 금액 나머지 존재")
     public void money4() {
         assertThatThrownBy(() -> {
-            new Money(1550);
+            new Money(BigDecimal.valueOf(1550));
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("[ERROR]");
     }
